requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
Bootstrap-Flask==2.2.0
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
colorgram.py==1.2.0
cssselect==1.3.0
decorator==5.2.1
defusedxml==0.7.1
discord.py==2.4.0
django-timedeltafield==0.7.10
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
executing==2.2.0
fastjsonschema==2.21.1
Flask==2.3.2
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
greenlet==3.1.1
gunicorn==22.0.0
idna==3.10
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
multidict==6.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.2.5
p3==0.2.1
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.4.13
platformdirs==4.3.7
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pygame==2.6.1
Pygments==2.19.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
python-dateutil==2.9.0.post0
pytz==2025.1
#pywin32==310
pyzmq==26.4.0
referencing==0.36.2
requests~=2.32.0
rpds-py==0.24.0
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.25
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.0
Wikipedia-API==0.7.1
WTForms==3.0.1
yarg==0.1.9
yarl==1.10.0
stripe~=12.3.0
zyphra~=0.1.6
PyMuPDF~=1.25.5
scikit-learn~=1.6.1
gspread~=6.2.1
oauth2client~=4.1.3